refactor(auth): split auth pages under max-lines-per-function (#196)
Refactor the three flagged auth components to under 150 lines each by
extracting their submit/validation logic into custom hooks and pulling
presentational sections into co-located sub-components. Pure,
behavior-preserving refactor — no UI, logic, or export-signature changes.

- signin/page.tsx: extract useSignIn hook + SignInMobileNav, SignInIllustration
- signup/SignUpForm.tsx: extract useSignUp hook + SignUpMobileNav, SignUpIllustration
- reset-password/page.tsx: extract useResetPassword hook + ResetPasswordMobileNav,
  ResetPasswordIllustration
- add shared AuthPulseDots component for the identical decorative dots
