+++ app/auth/signin/page.tsx
@@ -1,161 +1,25 @@
 "use client";
-import { useState, useRef, useEffect } from "react";
 import Image from "next/image";
-import { signIn, useSession } from "next-auth/react";
-import { useRouter, usePathname } from "next/navigation";
-import { z } from "zod";
-import { toast } from "sonner";
-
-const signInSchema = z.object({
-  email: z.string().min(1, "Please enter your email").email("Invalid email address"),
-  password: z.string().min(1, "Please enter your password"),
-});
+import { signIn } from "next-auth/react";
+import { useSignIn } from "./useSignIn";
+import SignInMobileNav from "./SignInMobileNav";
+import SignInIllustration from "./SignInIllustration";
 
 const SignIn = () => {
-  const [showPassword, setShowPassword] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-  const emailRef = useRef<HTMLInputElement>(null);
-  const passwordRef = useRef<HTMLInputElement>(null);
-  const [animatePanel, setAnimatePanel] = useState(false);
-  const router = useRouter();
-  const pathname = usePathname();
-  const { update } = useSession();
-
-  useEffect(() => {
-    const timeout = setTimeout(() => setAnimatePanel(true), 100);
-    return () => clearTimeout(timeout);
-  }, []);
-
-  const togglePasswordVisibility = () => setShowPassword(!showPassword);
-
-  // const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-  //   e.preventDefault();
-  //   const email = emailRef.current?.value ?? "";
-  //   const password = passwordRef.current?.value ?? "";
-  //   setIsLoading(true);
-  //   try {
-  //     signInSchema.parse({ email, password });
-  //   } catch (error) {
-  //     if (error instanceof z.ZodError) {
-  //       toast.error(error.errors[0].message);
-  //       setIsLoading(false);
-  //       return;
-  //     }
-  //   }
-  //   try {
-  //     const res = await signIn("credentials", {
-  //       email,
-  //       password,
-  //       redirect: false,
-  //     });
-  //     if (res?.ok) {
-  //       toast.success("Signed in successfully! from try");
-  //       console.log("Signed in successfully!");
-  //       console.log(status);
-  //       await update();
-  //       // await new Promise((resolve) => setTimeout(resolve, 1000));
-  //       router.push("/dashboard");
-  //       // setTimeout(() => window.location.reload(), 100);
-  //     } else {
-  //       toast.error(res?.error || "Invalid credentials.");
-  //     }
-  //   } catch (err) {
-  //     console.warn(err);
-  //     if (
-  //       err instanceof TypeError &&
-  //       err.message.includes("Failed to construct 'URL'")
-  //     ) {
-  //       toast.success("Signed in successfully! from catch");
-  //       console.log("signed in successfully from catch block");
-  //       await update();
-  //       // setTimeout(() => window.location.reload(), 100);
-  //       router.push("/dashboard");
-  //       return;
-  //     }
-  //     toast.error("Something went wrong. Please try again.");
-  //   } finally {
-  //     setIsLoading(false);
-  //   }
-  // };
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    const email = emailRef.current?.value ?? "";
-    const password = passwordRef.current?.value ?? "";
-    console.log(email, password, "first");
-    setIsLoading(true);
-
-    try {
-      // ✅ Validate input
-      signInSchema.parse({ email, password });
-      console.log(email, password);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        toast.error(error.errors[0].message);
-        setIsLoading(false);
-        return;
-      }
-    }
-
-    try {
-      // ✅ Sign in with credentials
-      const res = await signIn("credentials", {
-        email,
-        password,
-        redirect: false,
-      });
-      console.log(email, password);
-      console.log("Sign-in response:", res);
-
-      if (res?.ok) {
-        toast.success("Signed in successfully!");
-        await update(); // refresh session
-        router.push("/dashboard"); // ✅ only one redirect
-      } else {
-        toast.error(res?.error || "Invalid credentials.");
-      }
-    } catch (err) {
-      console.log("error from the new catch block");
-      console.error("Sign-in error:", err);
-      toast.error("Something went wrong. Please try again.");
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  const {
+    showPassword,
+    togglePasswordVisibility,
+    isLoading,
+    animatePanel,
+    emailRef,
+    passwordRef,
+    router,
+    handleSubmit,
+  } = useSignIn();
 
   return (
     <div className="flex flex-col md:flex-row min-h-screen w-full font-sans bg-[#F1ECFF]">
-      <div
-        className="md:hidden absolute top-0 left-0 w-full h-20 bg-gradient-to-b from-[#313053] to-[#261753] z-[1000] flex justify-center items-center shadow-lg"
-        style={{
-          borderBottomLeftRadius: "50% 20%",
-          borderBottomRightRadius: "50% 20%",
-        }}
-      >
-        <div className="flex bg-[#313053]/80 backdrop-blur-sm rounded-full p-1.5 shadow-inner">
-          <button
-            onClick={() => router.push("/auth/signin")}
-            className={`px-6 py-2 rounded-full transition-all duration-300 transform hover:scale-105 ${
-              pathname === "/auth/signin"
-                ? "bg-gradient-to-r from-[#615fa1] to-[#313053] text-white shadow-md"
-                : "text-gray-300 hover:bg-[#615fa1] hover:text-white"
-            }`}
-          >
-            Sign In
-          </button>
-          <button
-            onClick={() => router.push("/auth/signup")}
-            className={`px-6 py-2 rounded-full transition-all duration-300 transform hover:scale-105 ${
-              pathname === "/auth/signup"
-                ? "bg-gradient-to-r from-[#615fa1] to-[#313053] text-white shadow-md"
-                : "text-gray-300 hover:bg-[#615fa1] hover:text-white"
-            }`}
-          >
-            Sign Up
-          </button>
-        </div>
-      </div>
+      <SignInMobileNav />
 
       <div
         className={`w-full md:w-1/2 flex justify-center items-center px-4 sm:px-10 lg:px-20 py-6 md:py-10 transition-all duration-700 ease-out pt-24 md:pt-10 ${
@@ -258,38 +122,7 @@ const SignIn = () => {
           </div>
         </form>
       </div>
-      <div className="hidden md:flex md:w-1/2 relative justify-center items-center overflow-hidden rounded-l-[75px] bg-[#B09EE4]">
-        <div
-          className={`absolute inset-0 bg-[#261753] rounded-l-[75px] z-0 transition-all duration-700 ease-out ${
-            animatePanel ? "ml-[20px]" : "ml-[100%]"
-          }`}
-        />
-        <div className="relative z-10 px-6 sm:px-8">
-          <Image
-            src="/icons/login-vector.svg"
-            alt="Login Illustration"
-            width={400}
-            height={400}
-            className="max-w-full h-auto"
-          />
-        </div>
-        <button
-          onClick={() => router.push("/")}
-          className="absolute top-4 sm:top-6 right-6 sm:right-10 flex items-center gap-2 sm:gap-3 z-10 cursor-pointer"
-        >
-          <Image src="/icons/logo.png" alt="Logo" width={28} height={28} />
-
-          <span className="text-base sm:text-lg font-extrabold tracking-wider text-[#B09EE4]">
-            MARVEDGE
-          </span>
-        </button>
-
-        <div className="absolute top-1/4 right-1/4 w-3 h-3 bg-white/20 rounded-full animate-pulse hover:scale-150 transition-transform duration-300"></div>
-        <div className="absolute bottom-1/3 left-1/4 w-4 h-4 bg-white/15 rounded-full animate-pulse delay-1000 hover:scale-150 transition-transform duration-300"></div>
-        <div className="absolute top-1/2 left-1/3 w-2 h-2 bg-white/25 rounded-full animate-pulse delay-500 hover:scale-150 transition-transform duration-300"></div>
-        <div className="absolute top-1/5 left-1/5 w-2.5 h-2.5 bg-white/20 rounded-full animate-pulse delay-200 hover:scale-150 transition-transform duration-300"></div>
-        <div className="absolute bottom-1/5 right-1/3 w-3.5 h-3.5 bg-white/15 rounded-full animate-pulse delay-1200 hover:scale-150 transition-transform duration-300"></div>
-      </div>
+      <SignInIllustration animatePanel={animatePanel} />
     </div>
   );
 };
